verification bases get copied to hosting provider on approval
