+++ apps/accounts/models/hosting/provider.py
@@ -12,6 +12,8 @@ from django.db import models
 from django.template.loader import render_to_string
 from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
+from django.utils.translation import pgettext_lazy
+from django.utils.safestring import mark_safe
 from django_countries.fields import CountryField
 from django_mysql.models import EnumField
 from guardian.shortcuts import get_users_with_perms
@@ -72,6 +74,55 @@ class Service(tag_models.TagBase):
         verbose_name = _("Service")
         verbose_name_plural = _("Services")
 
+
+class VerificationBasis(tag_models.TagBase):
+    """
+    A model representing reasons why a provider would be verified.
+    This includes things like
+
+        "We use 100% green energy from our own infrastructure.",
+        "We operate in a region that has a grid intensity of less than 20g CO2e/kWh or uses over 99% renewable power.",
+        "We directly pay for green energy to cover the non-green energy we use.",
+        "We purchase quality carbon offsets to cover the non-green energy we use.",
+        "We resell or actively use a provider that is already in the Green Web Dataset.",
+
+    A subclass of Taggit's 'TagBase' model.
+
+    """
+
+
+    # Annoyingly, the only way to override the max_length in taggit appears to be to copy and adjust
+    # these two field definitions wholesale: https://github.com/jazzband/django-taggit/issues/510
+    name = models.CharField(
+        verbose_name=pgettext_lazy("A tag name", "name"), unique=True, max_length=255
+    )
+
+
+    slug = models.SlugField(
+        verbose_name=pgettext_lazy("A tag slug", "slug"),
+        unique=True,
+        max_length=255,
+        allow_unicode=True,
+    )
+
+    required_evidence_link = models.URLField(
+        max_length=255, null=True, blank=True,
+        verbose_name="Required evidence link"
+    )
+
+    class Meta:
+        verbose_name = _("Basis for verification")
+        verbose_name_plural = _("Bases for verification")
+
+    @property
+    def label(self):
+        label = self.name
+        if self.required_evidence_link is not None and self.required_evidence_link.strip() != "":
+            label += f" (<a href='{self.required_evidence_link}' target='_blank'>see required evidence</a>)"
+        return mark_safe(label)
+
+
+
 class ProviderService(tag_models.TaggedItemBase):
     """
     The corresponding through model for linking a Provider to
@@ -84,6 +135,22 @@ class ProviderService(tag_models.TaggedItemBase):
     )
     tag = models.ForeignKey(
         Service,
+        related_name="%(app_label)s_%(class)s_items",
+        on_delete=models.CASCADE,
+    )
+
+class ProviderVerificationBasis(tag_models.TaggedItemBase):
+    """
+    The corresponding through model for linking a Provider to
+    a VerificationBasis as outlined above.
+    """
+
+    content_object = models.ForeignKey(
+        "Hostingprovider",
+        on_delete=models.CASCADE,
+    )
+    tag = models.ForeignKey(
+        VerificationBasis,
         related_name="%(app_label)s_%(class)s_items",
         on_delete=models.CASCADE,
     )
@@ -129,7 +196,15 @@ class Hostingprovider(models.Model):
         through=ProviderService,
         # related_name="services",
     )
+
+    verification_bases = TaggableManager(
+        verbose_name="Basis for verification",
+        blank=True,
+        through=ProviderVerificationBasis,
+    )
+
     # this should not be exposed publicly to end users.
+
     # It's for internal use
     staff_labels = TaggableManager(
         verbose_name="Staff labels",
